refactor de nuevas rutas
