+++ .env
@@ -1,3 +1,11 @@
 REACT_APP_BASE_URL=https://ongapi.alkemy.org/api
 
-REACT_APP_CONTACTS=/contacts
+REACT_APP_CONTACTS=/contacts 
+REACT_APP_TESTIMONIALS=/testimonials
+REACT_APP_ACTIVITIES=/activities
+REACT_APP_MEMBERS=/members
+REACT_APP_ORGANIZATION=/organization 
+REACT_APP_SLIDES=/slides 
+REACT_APP_CATEGORIES=/categories
+REACT_APP_NEWS=/news
+REACT_APP_USERS=/users
