+++ src/App.css
@@ -1,3 +1,8 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+}
+
 .App {
   text-align: center;
 }
@@ -30,9 +35,11 @@
   0% {
     transform: translateY(0);
   }
+
   50% {
     transform: translateY(10px)
   }
+
   100% {
     transform: translateY(0px)
   }
