+++ src/App.js
@@ -1,20 +1,18 @@
 import React from 'react';
-import logo from './logo.svg';
-import { Counter } from './features/counter/Counter';
 import './App.css';
-import {BrowserRouter, Route, Switch} from 'react-router-dom';
-import ActivitiesForm from './Components/Activities/ActivitiesForm';
-import CategoriesForm from './Components/Categories/CategoriesForm';
-import NewsForm from './Components/News/NewsForm';
-import SlidesForm from './Components/Slides/SlidesForm';
-import TestimonialForm from './Components/Testimonials/TestimonialsForm';
-import UserForm from './Components/Users/UsersForm';
-import SchoolCampaign from './Campaigns/School/SchoolCampaign';
-import ToysCampaign from './Campaigns/Toys/ToysCampaign';
-import MembersForm from './Components/Members/MembersForm';
-import ProjectsForm from './Components/Projects/ProjectsForm';
-import About from './about/About';
-import Organization from './Components/Organization/Organization';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+// import ActivitiesForm from './Components/Activities/ActivitiesForm';
+// import CategoriesForm from './Components/Categories/CategoriesForm';
+// import NewsForm from './Components/News/NewsForm';
+// import SlidesForm from './Components/Slides/SlidesForm';
+// import TestimonialForm from './Components/Testimonials/TestimonialsForm';
+// import UserForm from './Components/Users/UsersForm';
+// import SchoolCampaign from './Campaigns/School/SchoolCampaign';
+// import ToysCampaign from './Campaigns/Toys/ToysCampaign';
+// import MembersForm from './Components/Members/MembersForm';
+// import ProjectsForm from './Components/Projects/ProjectsForm';
+import About from './Components/Frontoffice/About/About'
+import Organization from './Components/Backoffice/Organization/Organization';
 
 function App() {
   return (
@@ -22,7 +20,7 @@ function App() {
       <BrowserRouter>
         <Switch>
           {/* <Route path="/" exact component={} />           Esta ruta debe ser para el Home */}
-          <Route path="/create-activity" component={ActivitiesForm} />
+          {/* <Route path="/create-activity" component={ActivitiesForm} />
           <Route path="/create-category" component={CategoriesForm} />
           <Route path="/create-news" component={NewsForm} />
           <Route path="/backoffice/create-slide" component={SlidesForm} />
@@ -31,9 +29,9 @@ function App() {
           <Route path="/create-member" component={MembersForm} />
           <Route path="/create-project" component={ProjectsForm} />
           <Route path="/school-campaign" component={SchoolCampaign} />
-          <Route path="/toys-campaign" component={ToysCampaign} />
+          <Route path="/toys-campaign" component={ToysCampaign} /> */}
           <Route path="/about" component={About} />
-          <Route path="/backoffice/organization" component={Organization} />
+          <Route exact path="/backoffice/organization" component={Organization} />
         </Switch>
       </BrowserRouter>
     </>
