+++ src/Components/Backoffice/Header/Header.css
@@ -1,0 +1,20 @@
+.header-container {
+    padding: 10px;
+}
+
+.header-container>svg[data-testid="MenuIcon"] {
+    visibility: hidden;
+}
+
+@media only screen and (max-width: 600px) {
+    .header-container {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        padding: 10px 0 10px 10px;
+    }
+
+    .header-container>svg[data-testid="MenuIcon"] {
+        visibility: visible;
+    }
+}
