+++ src/Components/Backoffice/Organization/Organization.css
@@ -1,0 +1,32 @@
+.organization-container {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+
+    width: 100vw;
+    margin-top: 2rem;
+}
+
+.organization-container__div {
+    display: flex;
+    width: 75%;
+    justify-content: space-around;
+    align-items: center;
+}
+
+.organization-container__p {
+    width: 75%;
+    font-size: 1.5rem;
+    margin: 0 2rem;
+    font-family: 'Montserrat';
+    font-weight: bold;
+}
+
+.organization-container__div-img {
+    width : 30%;
+}
+
+.organization-container__button{
+    margin: 5rem auto;
+    width: 80%;
+}
