+++ src/Components/Cards/Slides/Slides.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Slides = () => {
+  return (
+    <div>Slides</div>
+  )
+}
+
+export default Slides
