+++ src/Services/getPrivateHeaderAuth.js
@@ -1,11 +1,10 @@
+import getToken from './getToken';
+
 const getPrivateHeaderAuth = () => {
     //Devuelve un objeto header con la autorizacion si es que hay un token valido
-    token = getToken();
-    header = {};
-    if(token){
-        header = {'Authorization': 'Bearer ' + token};
-    }
-    return header;
+    const token = getToken();
+
+    return token ? {'Authorization': 'Bearer ' + token} : {error: "No token found"};
 }
 
 export default getPrivateHeaderAuth
