+++ src/features/Spinner/Spinner.css
@@ -1,0 +1,15 @@
+.SpinnerDiv{
+    width: 100%;
+    min-height: 100vh;
+    background-color: black;
+    opacity: 0.75;
+    z-index: 5;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    position: absolute;
+    top:0;
+    left: 0;
+}
